+++ src/components/authentication/Login.vue
@@ -29,27 +29,27 @@
             <div>
               <div v-if="error" class="alert alert-danger"><p>{{error}}</p></div>
               <form action="#" @submit.prevent="submit">
-                <div class="form-group row">
-                  <label for="email" class="col-md-4 col-form-label text-md-right">
-                    <p>Email:</p>
-                  </label>
-
-                  <div class="col-md-6">
-                    <input
-                      id="email"
-                      type="email"
-                      class="form-control"
-                      name="email"
-                      placeholder="Email"
-                      autocomplete="email"
-                      value
-                      required
-                      autofocus
-                      v-model="form.email"
-                    />
-                  </div>
+               <div>
+                <label for="name" class="col-md-4 col-form-label text-md-right">
+                  <p class="fieldName">
+                    Knighthood Name:
+                  </p>
+                </label>
+                <div class="col-md-6">
+                  <input
+                    id="name"
+                    type="name"
+                    class="form-control"
+                    name="name"
+                    placeholder="Your Username"
+                    autocomplete="nickname"
+                    value
+                    required
+                    autofocus
+                    v-model="form.name"
+                  />
                 </div>
-
+                  </div>
                 <div class="form-group row">
                   <label for="password" class="col-md-4 col-form-label text-md-right">
                     <p>
@@ -135,7 +135,7 @@ export default {
   submit() {
       firebase
         .auth()
-        .signInWithEmailAndPassword(this.form.email, this.form.password)
+        .signInWithEmailAndPassword(this.form.name.replace(/\s/g, "")+'@fakeemail.com', this.form.password)
         .catch(err => {
           this.error = err.message;
         });
@@ -219,7 +219,7 @@ hr {
 }
 
 .container { 
-  max-width:450px;
+  max-width:600px;
 }
 
 button {