+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # V.1.3.3 - Roadblocks
 *Deployed 12/24/2020*
 - prevented access to game via mobile devices
+- for people who don't want to sign up with an email address I've allowed signup through a username and password
 
 
 # V.1.3.2 - More Tooltips & Bug Squashing